requirements.txt
pytest==8.3.3
python_pptx==1.0.2
